refactor(tasks): move Go2 joystick owner [skip ci]
