+++ src/unilab/tasks/locomotion/a2/joystick.py
@@ -28,7 +28,7 @@ from unilab.envs.locomotion.common.commands import sample_commands_with_standing
 from unilab.envs.locomotion.common.dr_provider import LocomotionDRProvider
 from unilab.envs.locomotion.common.rewards import RewardContext
 from unilab.envs.locomotion.go2.base import Asset, ControlConfig
-from unilab.envs.locomotion.go2.joystick import (
+from unilab.tasks.locomotion.go2.joystick import (
     Go2DomainRandConfig,
     Go2JoystickCfg,
     Go2JoystickDomainRandomizationProvider,
