+++ src/unilab/tasks/locomotion/go2/__init__.py
@@ -1,9 +1,12 @@
 from .footstand import Go2FootStandCfg, Go2FootStandTask
+from .joystick import Go2JoystickCfg, Go2WalkTask
 from .rough import Go2JoystickRoughCfg, Go2JoystickRoughEnv
 
 __all__ = [
     "Go2FootStandCfg",
     "Go2FootStandTask",
+    "Go2JoystickCfg",
     "Go2JoystickRoughCfg",
     "Go2JoystickRoughEnv",
+    "Go2WalkTask",
 ]
