+++ tests/tasks/test_package_boundary.py
@@ -13,9 +13,7 @@ _ENV_PACKAGE = _REPO_ROOT / "src" / "unilab" / "envs"
 
 _TASK_REGISTRY_MODULES = (
     "unilab.envs.locomotion.go1",
-    "unilab.tasks.locomotion.go2.footstand",
-    "unilab.envs.locomotion.go2",
-    "unilab.tasks.locomotion.go2.rough",
+    "unilab.tasks.locomotion.go2",
     "unilab.envs.locomotion.go2w",
     "unilab.envs.locomotion.g1",
     "unilab.envs.locomotion.go2_arm",
